+++ src/components/Card.ts
@@ -1,0 +1,5 @@
+import styled from 'styled-components';
+
+export const Card = styled.div`
+
+`
